+++ src/rer/newsletter/browser/channel/deleteexpiredusers.py
@@ -19,41 +19,43 @@ class DeleteExpiredUsersView(BrowserView):
         self.request = request
         self.user_removed = 0
 
-    def update_annotations(self, channel):
+    def update_annotations(self, brain):
+        channel = brain.getObject()
         expired_date = datetime.now()
         expired_time_token = self.context.portal_registry.get(
-            'rer.newsletter.browser.settings.ISettingsSchema.expired_time_token',  # noqa
+            "rer.newsletter.browser.settings.ISettingsSchema.expired_time_token",  # noqa
             None,
         )
 
         adapter = getMultiAdapter(
             (channel, self.request), IChannelSubscriptions
         )
-
-        for key, subscription in adapter.channel_subscriptions.items():
+        keys = [x for x in adapter.channel_subscriptions.keys()]
+        for key in keys:
+            subscription = adapter.channel_subscriptions[key]
             creation_date = datetime.strptime(
-                subscription['creation_date'], '%d/%m/%Y %H:%M:%S'
+                subscription["creation_date"], "%d/%m/%Y %H:%M:%S"
             )
             if (
                 creation_date + timedelta(hours=expired_time_token)
                 < expired_date  # noqa
-                and not subscription['is_active']  # noqa
+                and not subscription["is_active"]  # noqa
             ):
-                del subscription
+                del adapter.channel_subscriptions[key]
                 self.user_removed += 1
 
     def __call__(self):
         alsoProvides(self.request, IDisableCSRFProtection)
-        logger.info(u'START:Remove expired user from channels')
+        logger.info(u"START:Remove expired user from channels")
 
-        pc = getToolByName(self.context, 'portal_catalog')  # noqa
+        pc = getToolByName(self.context, "portal_catalog")  # noqa
         channels_brain = pc.unrestrictedSearchResults(
-            {'portal_type': 'Channel'}
+            {"portal_type": "Channel"}
         )
 
         list(map(lambda x: self.update_annotations(x), channels_brain))
         logger.info(
-            u'DONE:Remove {0} expired user from channels'.format(
+            u"DONE:Remove {0} expired user from channels".format(
                 self.user_removed
             )
         )